pyproject.toml
[project]
name = "ai-frontier"
version = "0.1.0"
description = "Automated arxiv paper translation and summarization reporting service"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "tqdm>=4.65.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.28.0",
    "arxiv==2.1.0",
    "feedparser>=6.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "schedule>=1.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "email-validator>=2.0.0",
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.7.4",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true